requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.11.0
openai>=1.50.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
